rename: 타입 및 파일명 변경
