+++ src/components/common/Header/HeaderRightCol.tsx
@@ -1,5 +1,5 @@
 import { formatDate } from "@utils/date/formatDate";
-import { ManagementVariant } from "@types/main";
+import { ManagementVariant } from "@types/member";
 import styled from "@emotion/styled";
 import { Button, Stack, Box, Typography } from "@mui/material";
 import { ReactElement } from "react";
